Prepare for transition to hand-tuned autostacking
Taskify autostacking functions, add option for hand-tuned or algorithmic
positioning of lift during autostacking, fix delays inside tasks.

+++ include/motorGroupControl.h
@@ -3,7 +3,7 @@
 
 #include "config.h"
 
-void waitForMovementToFinish(bool waitForChain=true, bool waitForLift=true, int timeout=75, double chainMargin=7, double liftMargin=10);
+void waitForMovementToFinish(bool waitForChain=true, bool waitForLift=true, unsigned short timeout=75, double chainMargin=7, double liftMargin=10);
 
 //#region lift
 void setLiftState(liftState state);
